# services/auth-service/src/main/java/dev/sagelms/auth/dto/LoginRequest.java
package dev.sagelms.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String password
) {}