# services/auth-service/src/main/java/dev/sagelms/auth/dto/RegisterRequest.java
package dev.sagelms.auth.dto;

import dev.sagelms.auth.entity.UserRole;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 8) String password,
        @NotBlank @Size(max = 255) String fullName,
        @NotNull UserRole role
) {}